Few fixes, as the last version didn't work
Added a test file, all it is to test is if I have it setup correctly as
node cannot run the ecma6.  Expected output is an error on the keyword
class.
Added an ApiResolver to figure out which environment we are in.
Currently only works in Firefox and will someday likely need to be in
its own repo
Removed the FirefoxInit as it was unnecessary.